refactor(gateway): cleanup
