+++ services/gateway/web/embeddable/templates/error.html
@@ -54,7 +54,7 @@
         }
 
         .container_text {
-            font-family: 'Latio';
+            font-family: 'Lato';
             text-align: center;
         }
