+++ services/gateway/web/embeddable/templates/installation.html
@@ -53,7 +53,7 @@
         }
 
         .container_text {
-            font-family: 'Latio';
+            font-family: 'Lato';
             text-align: center;
         }
